+++ CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.0] - 2022-07-02
+### Added
+- MOM using RabbitMQ
+- Nodes implementation
+- Filters and transforms
+- Unreplicated coordinator with docker-in-docker, monitoring heartbeats
+- Dockerized client and coordinator in compose
+- Settings for an initial version of average posts score
+
+### Modified
+- Client now uses MOM
+- RabbitMQ now gets configuration from file in volume
+- Sidecars now use threads instead of processes
+
 ## [0.5.0] - 2022-07-01
 ### Added
 - Client that writes straight to rabbit